dbのAPI化
